rename: components 폴더에 Post 폴더 추가
Post 관련 컴포넌트를 체계적으로 관리하기 위해 Post 폴더 추가
